+++ ArtOfIllusion/src/artofillusion/object/Cylinder.java
@@ -457,7 +457,7 @@ public class Cylinder extends Object3D
   @Override
   public void setPropertyValue(int index, Object value)
   {
-    double val = ((Double) value).doubleValue();
+    double val = ((Double) value);
     if (index == 0)
       setSize(2.0*val, height, 2.0*rz);
     else if (index == 1)
